Initial code drop
A small library of various GeoCoordinate functionality. Including
defining a basic type, ability to perform a geohash on it, and do some
proximity search.

Structually this is built as an internal library with a public one
that presents a limited interface. This allows for tests to have the
items required for them, while limiting the externally exposed
interface. Notably some typeclass instances are limited to tests as
well as allowing constructors to be used internally (again for
typeclass instance derivation), but not exposed externally.